Implemented process style systems
Process style systems use a const member function taking all components it wants as const-qualified parameters. All state of the class must not be modified. While const guarantees most of this, const-casting it away or modifying mutable data members will cause race conditions.

+++ src/Simulation.cpp
@@ -4,7 +4,7 @@
 
 
 void Simulation::Run( double timestep, const std::string &world ) {
-  this->GetWorld( world ).Update( timestep );
+  this->GetWorld( world ).Update( (float)timestep );
 }
 World& Simulation::CreateWorld( const std::string &name ) {
   auto result{ m_Worlds.emplace( name, name ) };
